Move stubs to separate file
